Goldfish simulator framework
