+++ src/main/java/org/mtgpeasant/perfectdeck/common/cards/Deck.java
@@ -1,0 +1,11 @@
+package org.mtgpeasant.perfectdeck.common.cards;
+
+import lombok.Builder;
+import lombok.Value;
+
+@Builder
+@Value
+public class Deck {
+    final Cards main;
+    final Cards sideboard;
+}
