+++ src/main/java/org/mtgpeasant/perfectdeck/common/matchers/MatcherContext.java
@@ -1,0 +1,5 @@
+package org.mtgpeasant.perfectdeck.common.matchers;
+
+public interface MatcherContext {
+    Matcher findByName(String name);
+}
