+++ src/main/java/org/mtgpeasant/perfectdeck/common/matchers/NoopMatcher.java
@@ -1,0 +1,24 @@
+package org.mtgpeasant.perfectdeck.common.matchers;
+
+import lombok.Builder;
+import lombok.Value;
+
+import java.util.stream.Stream;
+
+@Builder
+@Value
+class NoopMatcher implements Matcher {
+    @Override
+    public String toString() {
+        return "$noop";
+    }
+
+    @Override
+    public void validate(Validation validation, MatcherContext context) {
+    }
+
+    @Override
+    public Stream<Match> matches(Stream<Match> stream, MatcherContext context) {
+        return stream;
+    }
+}
