+++ src/main/java/org/mtgpeasant/perfectdeck/goldfish/GameLostException.java
@@ -1,0 +1,7 @@
+package org.mtgpeasant.perfectdeck.goldfish;
+
+public class GameLostException extends RuntimeException {
+    public GameLostException(String message) {
+        super(message);
+    }
+}
