+++ src/main/java/org/mtgpeasant/perfectdeck/goldfish/IllegalMoveException.java
@@ -1,0 +1,7 @@
+package org.mtgpeasant.perfectdeck.goldfish;
+
+public class IllegalMoveException extends RuntimeException {
+    public IllegalMoveException(String message) {
+        super(message);
+    }
+}
