--- src/main/java/org/mtgpeasant/stats/cards/Deck.java
@@ -1,11 +1,0 @@
-package org.mtgpeasant.stats.cards;
-
-import lombok.Builder;
-import lombok.Value;
-
-@Builder
-@Value
-public class Deck {
-    final Cards main;
-    final Cards sideboard;
-}
