--- src/main/java/org/mtgpeasant/stats/matchers/MatcherContext.java
@@ -1,5 +1,0 @@
-package org.mtgpeasant.stats.matchers;
-
-public interface MatcherContext {
-    Matcher findByName(String name);
-}
